Made things compile under SunOS4 again. Fix lots of warnings under SunOS4.
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@1201 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/sl/sl.c
@@ -135,7 +135,7 @@ sl_loop (SL_cmd *cmds, char *prompt)
     ptr = malloc(max_count * sizeof(*ptr));
     if (ptr == NULL) {
 	printf ("sl_loop: failed to allocate %u bytes of memory\n",
-		max_count * sizeof(*ptr));
+		(int) max_count * sizeof(*ptr));
 	return -1;
     }
 
@@ -164,7 +164,7 @@ sl_loop (SL_cmd *cmds, char *prompt)
 		    if (ptr == NULL) {
 			printf ("sl_loop: failed to allocate %u "
 				"bytes of memory\n",
-				max_count * sizeof(*ptr));
+				(int) max_count * sizeof(*ptr));
 			return -1;
 		    }
 		}
